% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockPredictor.R
\name{predict_stock}
\alias{predict_stock}
\title{Predict Stock Price}
\usage{
predict_stock(symbol)
}
\arguments{
\item{symbol}{}
}
\description{
This package help you predics whether the stock price would be higher or lower at tomorrow's closing compared to today's closing.
}
\examples{
predict_stock('MSFT')

}
